Include actual region from event for embedded link
Glad to have come across this template. However, for the notification portion of a failed pipeline, the region was hardcoded to us-east-1. Considering the template has Ref calls to the AWS Region, I would assume this is not desired behavior. According to https://docs.amazonaws.cn/en_us/AmazonCloudWatch/latest/events/EventTypes.html#codepipeline_event_type we can grab the region for the pipeline and include it in the embedded url.